refactor(GameResult): 승패 결정 기능을 GameResult가 하도록 변경
