+++ src/main/java/model/result/GameResult.java
@@ -1,5 +1,7 @@
 package model.result;
 
+import model.cards.Cards;
+
 public enum GameResult {
     WIN("승"),
     LOSE("패"),
@@ -17,6 +19,26 @@ public enum GameResult {
         }
         if (LOSE.equals(gameResult)) {
             return WIN;
+        }
+        return DRAW;
+    }
+
+    public static GameResult determineGameResult(final Cards dealerCards, final Cards playerCards) {
+        if (playerCards.isBust()) {
+            return GameResult.LOSE;
+        }
+        if (dealerCards.isBust()) {
+            return GameResult.WIN;
+        }
+        return determineGameResultByScore(dealerCards, playerCards);
+    }
+
+    private static GameResult determineGameResultByScore(final Cards dealerCards, final Cards playerCards) {
+        if (playerCards.calculateResult() > dealerCards.calculateResult()) {
+            return GameResult.WIN;
+        }
+        if (playerCards.calculateResult() < dealerCards.calculateResult()) {
+            return GameResult.LOSE;
         }
         return DRAW;
     }
